refactor: create footer utils folder
